Task done
This assignment implemented a small blog using the jsonplaceholder API.

The initial page displays information about all users, below the page there is a news line with random user posts.

Clicking on a user takes you to the biography of the selected user and links to their posts.

After selecting a post, you go to the page with the post and comments to it.

When selecting any post from the news feed, the user is redirected to the page with the selected post.

After any post has been selected, the user can click on the author's name and it will be taken to a page with full information about the author.

At any time, the user can click on the "Home" button and return to the main page.